algostack/pyproject.toml
[tool.poetry]
name = "algostack"
version = "0.1.0"
description = "Disciplined multi-strategy algorithmic trading stack"
authors = ["AlgoStack Team"]
readme = "README.md"
packages = [
    {include = "algostack"},
    {include = "adapters"},
    {include = "api"},
    {include = "backtests"},
    {include = "cli"},
    {include = "core"},
    {include = "strategies"},
    {include = "utils"}
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
numpy = "^1.24.0"
scipy = "^1.10.0"
yfinance = "^0.2.28"
requests = "^2.31.0"
ccxt = "^4.0.0"
TA-Lib = "^0.4.28"
statsmodels = "^0.14.0"
backtrader = "^1.9.78.123"
ib-insync = "^0.9.86"
robin-stocks = "^3.0.8"
pydantic = "^2.0.0"
PyYAML = "^6.0"
apscheduler = ">=3.10.4"
fastapi = "^0.103.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
prometheus-client = "^0.17.0"
alpha-vantage = "^2.3.1"
pyarrow = "^14.0.0"
click = "^8.1.0"
scikit-learn = ">=1.3.0"
scikit-optimize = ">=0.9.0"
aiohttp = ">=3.9.0"
websockets = ">=12.0"
rich = ">=13.7.0"
fastparquet = ">=2024.0.0"
ib-async = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
hypothesis = "^6.82.0"
black = "^23.7.0"
ruff = "^0.0.287"
mypy = "^1.5.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.16.0"
freezegun = "^1.5.2"
psutil = "^7.0.0"
optuna = "^4.3.0"
pytest-benchmark = "4.0.0"
toml = "^0.10.2"
pytest-mock = ">=3.11.0"
httpx = ">=0.24.0"
faker = ">=19.0.0"
streamlit = ">=1.29.0"
pre-commit = "^4.2.0"
coverage = "^7.8.2"

[tool.poetry.group.optional.dependencies]
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

# Markers for test categories
markers = [
    "unit: Fast unit tests (deselect with '-m \"not unit\"')",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Tests that take > 1 second",
    "requires_market_data: Tests that need market data",
    "asyncio: mark test as async",
    "benchmark: Performance benchmark tests",
]

# Coverage settings
addopts = """
    -v
    --strict-markers
    --tb=short
"""

pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Temporarily disabled for test infrastructure setup:
# --cov=algostack
# --cov-report=html
# --cov-report=term-missing
# --cov-fail-under=95

# Test discovery
norecursedirs = [
    ".git",
    ".tox", 
    "dist",
    "build",
    "*.egg",
]

[tool.coverage.run]
source = ["algostack"]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 80